local models: move backend classification into modelBackend.ts
