+++ packages/agency-lang/lib/stdlib/localModels.ts
@@ -23,6 +23,21 @@ import {
 import { __ctx } from "../runtime/asyncContext.js";
 import { recordDownload, readDownloadManifest } from "./localModelManifest.js";
 import {
+  type Backend,
+  isGgufPath,
+  isMlxUri,
+  parseMlxUri,
+  isModelDir,
+  backendOfTarget,
+} from "./modelBackend.js";
+export {
+  type Backend,
+  isMlxUri,
+  parseMlxUri,
+  isModelDir,
+  backendOfTarget,
+} from "./modelBackend.js";
+import {
   MLX_SUBDIR,
   mlxModelDir,
   mlxModelDirName,
@@ -30,10 +45,6 @@ import {
   isMlxModelComplete,
 } from "./mlxModelRecord.js";
 import { ttyColor } from "../utils/termcolors.js";
-
-/** Which engine runs a model. GGUF files run in-process through llama.cpp.
- *  MLX models run in a server started with `agency local serve`. */
-export type Backend = "llama-cpp" | "mlx";
 
 /** What a model is FOR — a single axis, orthogonal to size (size is conveyed
  *  by `params` / `sizeBytes`). Lets the CLI group + filter without parsing the
@@ -382,59 +393,8 @@ export function _modelsCacheDir(cacheDir: string = ""): string {
   return resolveCacheDir(cacheDir);
 }
 
-function isGgufPath(v: string): boolean {
-  return v.endsWith(".gguf");
-}
-
 function isModelUri(v: string): boolean {
   return /^(hf:|https?:|mlx:)/.test(v);
-}
-
-/** `mlx:<org>/<repo>` with an optional `@<revision>`. */
-const MLX_URI = /^mlx:([^@\s/]+\/[^@\s/]+)(?:@([\w.-]+))?$/;
-
-export function isMlxUri(v: string): boolean {
-  return MLX_URI.test(v);
-}
-
-/** Split "mlx:org/repo@rev" into its parts. Throws on any other shape. */
-export function parseMlxUri(v: string): { repo: string; revision: string | undefined } {
-  const m = MLX_URI.exec(v);
-  if (m === null) {
-    throw new Error(
-      `"${v}" is not an mlx: URI. Expected mlx:<org>/<repo> or mlx:<org>/<repo>@<revision>.`,
-    );
-  }
-  return { repo: m[1], revision: m[2] };
-}
-
-/** A Hugging Face model directory: config.json plus at least one weights
- *  file. This is the layout `mlx_lm.server` loads. A GGUF model is one file
- *  with that information inside it, so it never has a config.json. */
-export function isModelDir(p: string): boolean {
-  const located = wholePath(p);
-  const info = stat(located.root, located.target);
-  if (info === null || !info.isDirectory()) {
-    return false;
-  }
-  const entries = list(root(p), ".");
-  const hasConfig = entries.some((e) => e.type === "file" && e.name === "config.json");
-  const hasWeights = entries.some((e) => e.type === "file" && e.name.endsWith(".safetensors"));
-  return hasConfig && hasWeights;
-}
-
-/** Which engine runs a resolved target. Throws for a path that is neither a
- *  GGUF file nor a model directory. */
-export function backendOfTarget(target: string): Backend {
-  if (isGgufPath(target) || /^(hf:|https?:)/.test(target)) {
-    return "llama-cpp";
-  }
-  if (isMlxUri(target) || isModelDir(target)) {
-    return "mlx";
-  }
-  throw new Error(
-    `"${target}" is not a model: expected a .gguf file or a directory containing config.json.`,
-  );
 }
 
 /** A model URI we'll accept from the *remote catalog*. Stricter than
